feat: #123 최상위 폴더 삭제 방지 - parentFolder가 null일 경우, 삭제 방지 - 최상위 폴더 테스트 코드 추가 - mockRootFolder, mockSubFolder로 나누어 최상위, 하위 폴더 구분

+++ src/main/java/com/edio/studywithcard/folder/service/FolderServiceImpl.java
@@ -138,6 +138,9 @@ public class FolderServiceImpl implements FolderService {
     public void deleteFolder(Long folderId) {
         Folder existingFolder = folderRepository.findByIdAndIsDeletedFalse(folderId).get();
 
+        if(existingFolder.getParentFolder() == null){
+            throw new IllegalArgumentException(ErrorMessages.BAD_REQUEST.getMessage());
+        }
         folderRepository.delete(existingFolder);
     }
 }
